backend/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
chromadb
openai
pypdf
python-dotenv
fastapi
uvicorn[standard]
chardet